Add reindex step to bigquery workflow (#37)
Closes #38 

This PR adds the "reindexing" step from the `fetch` workflow to the `bigquery` workflow. This ensures that after we fetch all plugin versions from bigquery, our `extended_summary.json` (queried by `napari-plugin-manager`) is updated to reflect all available plugin versions. This means if a plugin is updated, we'll see its latest version even if our manifest fetching fails (or doesn't run, which is currently the case).

**NOTE**: This PR **does not** fetch manifests for brand new plugins. New plugins will not be listed, because their manifests are never fetched. I will open an issue and separate PR for this. 

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>