Remove summary endpoint in favour of extended summary (#51)
As discussed in #48, nobody is using the `summary.json` endpoint, and the `extended_summary.json` contains all the keys in `summary.json` + the conda and pypi versions for a package. 

Ahead of improving the name normalization aspect, also discussed in #48, this PR removes the `summary` endpoint and keeps only `extended_summary`.

---------

Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>
Co-authored-by: Carol Willing <carolcode@willingconsulting.com>