Use normalized package names where suitable (#52)
Closes #48 

This PR updates npe2api to use the **normalized** plugin name as the plugin key, and adds a `name` key to dictionaries holding the original, non-normalized name. Additionally, it ensures that all individual plugin files in the `public/pypi`, `public/conda` and `public/manifest` directories are named by their normalized name. 

To simplify retrieval of the manifests via the API, we normalize the name in the dynamic routing handler, so that the api endpoint for the manifest still uses the original name.