Delete unused files (#57)
In #56 we added `.gitignore` files to the three directories we were emptying to ensure the directories themselves were not deleted (as git doesn't distinguish a directory as an object). After #51, the `public/summary.json` file should have been deleted, but due to the scheduled workflow running and us not realizing it before merge, the file itself persisted.

This PR deletes these four files as they are no longer needed or used.