Add pypi endpoint to fetch individual plugin pypi info (#58)
In looking to fix the issue with release dates in napari/hub-lite#18, I realized we'd need to grab the `pypi` file for each plugin, as it contains the dates for each release. 

The `conda` file has dates for each release too, but not all plugins have conda, so I think it makes sense to fetch this info from pypi at first.

This PR adds a `pypi/<plugin-name>` endpoint, so that we can access the pypi release information.