Add Code of Conduct section to README (#67)
Adds code of conduct section (taken from `napari/napari`) to this repo.
As well as being good practice (and something we should probably do for
all org repos), this is in preparation for asking netlify for an [Open
Source Plan](https://www.netlify.com/legal/open-source-policy/) so that
we can move away from Vercel.

---------

Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>